ci: fix missing GITHUB_TOKEN reference (#18)
This pull request updates the CI workflow configuration to improve
security by replacing the use of environment variables with GitHub
secrets for token management.

Security improvements in CI workflow:

*
[`.github/workflows/ci-application.yml`](diffhunk://#diff-9af3aef6896bc42e496b26b00d5dc67d93b92ac8b1471cf541408a53ac7dfb2dL78-R80):
Replaced `env.GITHUB_TOKEN` with `secrets.GITHUB_TOKEN` for both the
`env` and `with` sections to ensure secure handling of the GitHub token.