ci: add conditional statement to run on release created (#22)
This pull request makes a small improvement to the continuous deployment
workflow by adding a conditional check to ensure that a release is
created before proceeding with the "Build and Push Docker Image" job.

*
[`.github/workflows/cd.yml`](diffhunk://#diff-ea3ea8c9932adc7ba8161ceda844fedd43b006848ef1140c050cbd7ea0788a18R45):
Added an `if` condition to the `jobs:` section to check if
`RELEASE_CREATED` is true before executing the job.